+++ demo/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>UI Managers</title>
+    <script src="http://localhost:8080/kaltura-ovp-player.js"></script>
+    <script src="./ui-manager.js"></script>
+  </head>
+  <body>
+    <h1>yonihhh</h1>
+  </body>
+</html>
