--- side-panels/demo/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Side Panels Manager</title>
-    <!--    <script src="https://cdnapisec.kaltura.com/p/2215841/embedPlaykitJs/uiconf_id/42237501/"></script>-->
-    <script src="http://localhost:8080/kaltura-ovp-player.js"></script>
-    <!--    <script src="https://unpkg.com/preact@10.6.5/dist/preact.umd.js"></script>-->
-    <script src="./playkit-panels-manager.js"></script>
-  </head>
-  <body>
-    <h1>yonihhh</h1>
-  </body>
-</html>
