--- side-panels/src/index.ts
@@ -1,10 +1,0 @@
-import {registerPlugin} from 'kaltura-player-js';
-import {pluginName, SidePanelsManagerService} from './side-panels-service';
-
-registerPlugin(pluginName, SidePanelsManagerService);
-
-export {SidePanelItem, SidePanelPosition, SidePanelMode, ReservedPresetName} from './side-panel-item';
-
-export interface jjj {
-    ffff: string;
-}
