+++ src/services/side-panels/side-panel-item.ts
